+++ src/app/commands/admin/GuestPass.ts
@@ -1,10 +1,7 @@
 import { SlashCommand, CommandOptionType, ApplicationCommandPermissionType, CommandContext } from 'slash-create';
-import db from '../../db/db';
-import constants from '../../constants';
-import { retrieveGuestRole } from '../../service/GuestPassService';
 import client from '../../app';
-
-const expiresInHours = 168;
+import roleIDs from '../../service/constants/roleIDs';
+import AddGuestPass, { expiresInHours } from '../../service/guest-pass/AddGuestPass';
 
 module.exports = class GuestPass extends SlashCommand {
 	constructor(creator) {
@@ -29,7 +26,12 @@ module.exports = class GuestPass extends SlashCommand {
 				[process.env.DISCORD_SERVER_ID]: [
 					{
 						type: ApplicationCommandPermissionType.ROLE,
-						id: process.env.DISCORD_ROLE_LEVEL_2,
+						id: roleIDs.level2,
+						permission: true,
+					},
+					{
+						type: ApplicationCommandPermissionType.ROLE,
+						id: roleIDs.admin,
 						permission: true,
 					},
 				],
@@ -39,93 +41,19 @@ module.exports = class GuestPass extends SlashCommand {
 	}
 
 	async run(ctx: CommandContext) {
-		// Ignores commands from bots
 		if (ctx.user.bot) return;
-
+		
+		console.log('/guest-pass start');
 		const guild = await client.guilds.fetch(ctx.guildID);
+		const guestUser = await guild.members.fetch(ctx.options.user);
 
-		// Guild member to assign guest pass role
-		const guildMember = await guild.members.fetch(ctx.options.user);
-
-		if (guildMember.user.bot) {
-			ctx.send('Bots don\'t need a guest pass!');
-			return;
+		if (guestUser.user.bot) {
+			return ctx.send('Bots don\'t need a guest pass!');
 		}
+		
+		await AddGuestPass(guestUser);
 
-		// Retrieve Guest Pass Role
-		const guestRole = await retrieveGuestRole(guild.roles);
-
-		// Open database connection
-		db.connect(constants.DB_NAME_DEGEN, async (err) => {
-			if (err) {
-				console.error('ERROR:', err);
-				return;
-			}
-			// DB Connected
-			const dbGuestUsers = await db.get().collection(constants.DB_COLLECTION_GUEST_USERS);
-			const queryOptions = {
-				upsert: true,
-			};
-			const currentTimestamp = Date.now();
-			const guestDbUser = {
-				_id: guildMember.id,
-				startTimestamp: currentTimestamp,
-				expiresTimestamp: currentTimestamp + (expiresInHours * 1000 * 60 * 60),
-			};
-
-			// Find and update guest user in mongodb
-			const dbUpdateResult = await dbGuestUsers.findOneAndReplace({
-				_id: guildMember.id,
-			}, guestDbUser, queryOptions);
-			if (dbUpdateResult == null) {
-				console.error('Failed to insert into DB');
-				return;
-			}
-
-			await db.close();
-			console.log(`user ${guildMember.id} inserted into guestUsers`);
-
-			// Add role to member
-			guildMember.roles.add(guestRole).catch(console.error);
-			console.log(`user ${guildMember.id} given ${constants.DISCORD_ROLE_GUEST_PASS} role`);
-			ctx.send(`Hey <@${guildMember.id}>! You now have access for ${expiresInHours / 24} days.`);
-		});
-
-		// Send out notification on timer
-		client.setTimeout(() => {
-			guildMember.send(`Hey <@${guildMember.id}>, your guest pass is set to expire in 1 day. Let us know if you have any questions!`);
-		}, (expiresInHours * 1000 * 60 * 60) - (1000 * 60 * 60 * 24));
-
-		client.setTimeout(() => {
-			guildMember.send(`Hey <@${guildMember.id}>, your guest pass is set to expire in 15 minutes. Let us know if you have any questions!`);
-		}, (expiresInHours * 1000 * 60 * 60) - (1000 * 60 * 15));
-
-		// Handle removal of guest pass
-		client.setTimeout(() => {
-			db.connect(constants.DB_NAME_DEGEN, async (err) => {
-				if (err) {
-					console.error('ERROR:', err);
-					return;
-				}
-				const dbGuestUsers = db.get().collection(constants.DB_COLLECTION_GUEST_USERS);
-				const guestDBQuery = {
-					_id: guildMember.id,
-				};
-				const dbDeleteResult = await dbGuestUsers.findOneAndDelete(guestDBQuery);
-				if (dbDeleteResult == null) {
-					console.error('Failed to remove from DB');
-					return;
-				}
-				await db.close();
-				console.log(`guest pass removed for ${guildMember.id} in db`);
-
-				// Remove guest pass role
-				guildMember.roles.remove(guestRole).catch(console.error);
-
-				console.log(`guest pass removed for ${guildMember.id} in discord`);
-
-				return guildMember.send(`Hi <@${guildMember.id}>, your guest pass has expired. Let us know at Bankless DAO if this was a mistake!`);
-			});
-		}, expiresInHours * 1000 * 60 * 60);
+		return ctx.send(`<@${ctx.user.id}> guest pass added and message sent!`);
 	}
 };
+