+++ src/app/commands/notion/NotionFAQs.ts
@@ -30,6 +30,7 @@ module.exports = class NotionFAQs extends SlashCommand {
 	async run(ctx) {
 		// Ignores commands from bots
 		if (ctx.user.bot) return;
+		console.log('/faqs start');
 
 		const guild = await client.guilds.fetch(ctx.guildID);
 		const guildMember = await guild.members.fetch(ctx.user.id);
@@ -102,11 +103,11 @@ module.exports = class NotionFAQs extends SlashCommand {
 	}
 };
 
-module.exports.retrieveFAQsPromise = async () => {
+module.exports.retrieveFAQsPromise = async (): Promise<Array<any>> => {
 	const faqs = [];
 	const numberRegex = /^[0-9]./;
 	const response = await notion.blocks.children.list({
-		block_id: process.env.FAQS_PAGE_ID
+		block_id: process.env.FAQS_PAGE_ID,
 	});
 	response.results.forEach((obj) => {
 		if (obj.type === 'paragraph' && obj.paragraph.text.length > 0) {