+++ src/app/events/ready.ts
@@ -2,19 +2,17 @@
  * Handler for Discord event `ready`.
  */
 
-import GuestPassService from '../service/GuestPassService';
-// import ScoapHttpServer from '../service/ScoapHttpServer';
+import GuestPassService from '../service/guest-pass/GuestPassService';
 import ScoapFastifyServer from '../service/ScoapFastifyServer';
 
 module.exports = {
 	name: 'ready',
 	once: true,
 
-	execute(client) {
+	async execute(client) {
 		console.log('Discord bot ready!');
 		client.user.setActivity('Going Bankless, Doing the DAO');
-		GuestPassService(client);
-		// ScoapHttpServer(client);
+		await GuestPassService(client);
 		ScoapFastifyServer();
 	},
 };