+++ src/test/service/bounty/create/new.test.ts
@@ -1,5 +1,0 @@
-describe('BountyNewTest', () => {
-	describe('todo', () => {
-		console.log('todo');
-	});
-});