+++ docs/CHANGELOG.md
@@ -3,6 +3,15 @@
 ## 1.2.0-SNAPSHOT
 
 1. Add Bounty slash command
+2. Add more unit cases for commands
+3. Add emoji support for bounties posted in #🧀-bounty-board channel
+   - restructure codebase
+   - add more logging
+4. Fix guest pass event service
+
+## 1.1.1-RELEASE (2021-07-28)
+
+1. Handle all notion api calls in case of rate limit or out of sync errors
 
 ## 1.1.0-RELEASE (2021-07-21)